requirements.txt
influxdb~=5.3.1
python_dateutil~=2.8.2
pyTibber~=0.27.1